requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.3.0
pydantic-settings==2.0.3

# LangChain & LangGraph
langchain==0.1.4
langchain-core==0.1.16
langchain-community==0.0.16
langgraph==0.0.20
langsmith==0.0.83

# LLM Providers
langchain-google-genai==0.0.6
langchain-openai==0.0.5
langchain-anthropic==0.0.1
google-generativeai==0.3.2

# Kubernetes
kubernetes==29.0.0
pyyaml==6.0.1

# AWS (for S3 kubeconfig)
boto3==1.34.34

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0